refactor(renderer): extract wikilink + link modules from RenderConfig
Moves the link-resolution helpers out of RenderConfig and into two new
crate-internal modules, leaving RenderConfig as a pure settings bag
(parser options + section/title-resolver storage).

- wikilink: WikilinkResolution enum + resolve(cfg, ...) +
  display_text(cfg, ...). Used by Walker on Tag::Link / WikiLink.
- link: strip_origin(cfg, ...) + section_ref_attrs(cfg, ...). Used by
  Walker on non-wikilink Tag::Link and Tag::Image.

Both modules expose free functions taking &RenderConfig as their first
arg. Tests move with their functions; no behavior change.

Seam 2 of crates/rw-renderer/REFACTOR.md.
